media-proxy/pyproject.toml
[project]
name = "media-proxy"
version = "0.1.0"
description = "simple media proxy for kaj.gg"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiohttp>=3.13.2",
  "python-dotenv>=1.2.1",
  "sanic>=25.3.0",
  "sanic-ext>=24.12.0",
]
